analytics/requirements.txt
Flask==2.3.2
pymongo==4.2.0
gunicorn==20.1.0
python-dotenv==0.19.0
Flask-PyMongo==2.3.0
Flask-CORS==3.0.10
openai==1.14.3
flask_graphql==2.0.0
graphql-core==2.3.2
graphql-relay==3.1.0
graphene==3.1
graphene-mongo==0.2.0
schema==0.7.4
graphql-core==2.3